env_requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==4.2.18
django-autocomplete-light==3.12.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-ninja==1.3.0
Faker==37.1.0
mysqlclient==2.2.6
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2